FIX: fetch-mock Response.body is not always a stream
The WhatWG fetch spec requires that Body.body (and therefore Response.body)
is always a readable stream, but it doesn't always happen that way in Node.
This seems to be because ResponseBuilder tries to fetch the stream module
from fetchMock, but it looks in the wrong place.

Fixes wheresrhys/fetch-mock#609